Big update. Lots of changs.
-created apiCalls.py
    -all functions for api calls in this file now
-created fileData.py
     -has function to read file
     -has class that stores video data in memory
-youtubeClickbait.py is now the main
-textAnalysis.py
    -started text analysis file
    -does a bad reg ex of all the titles
-updated youtubeVideos file
