+++ titlesTest.txt
@@ -1,0 +1,586 @@
+,000
+Player
+Building
+Challenge!
+Try
+Not
+To
+Laugh
+Challenge
+#82
+-
+Hard
+Mode!
+Doctor
+Reacts
+To
+The
+Worst
+TikTok
+Medical
+Advice
+We
+Improved
+Ocean
+Animals
+I'm
+pissed
+off.
+Korn
+-
+Start
+The
+Healing
+(Official
+Music
+Video)
+$1.2
+Trillion
+Bill
+Sent
+To
+The
+President's
+Desk
+|
+November
+,
+
+When
+a
+kid
+thinks
+he’s
+to
+smart
+for
+his
+grade
+:
+Let's
+unf**k
+the
+planet,
+one
+YouTube
+Short
+at
+a
+time
+#UnFckItChallenge
+#shorts
+DOWNTON
+ABBEY:
+A
+NEW
+ERA
+-
+Only
+in
+Theaters
+March
+
+NC
+teen
+girl
+rescued
+after
+flashing
+TikTok
+hand
+signals
+about
+danger
+What
+I
+Didn't
+Show
+You
+From
+The
+Twins
+Birth...
+Canelo
+Alvarez
+TKOs
+Caleb
+Plant
+in
+th
+Round
+|
+Highlights
+&
+Reaction
+|
+CBS
+Sports
+HQ
+I
+built
+
+mods
+that
+turn
+GTA
+Street
+Races
+into
+hell
+What's
+Left
+of
+Ed's
+LJ20
+That
+Was
+Stuck
+
+Years
+In
+The
+Sierra
+Nevada
+Mountains?
+Juice
+That
+Makes
+Your
+Head
+Explode
+The
+AFTERMATH
+of
+listing
+all
+
+cars
+for
+sale
+(most
+are
+GONE)
+WE
+SURPRISED
+RUBEN
+AND
+KAREN
+WITH
+THIS....😳
+Canelo
+Alvarez
+vs
+Caleb
+Plant
+-
+Film
+Study
+Season
+
+Trailer
+|
+JURASSIC
+WORLD
+CAMP
+CRETACEOUS
+|
+NETFLIX
+I
+Dry-Aged
+the
+WAGYU
+of
+the
+SWAMP
+and
+this
+happened!
+GTA
+
+Roleplay
+-
+SUPERCAR
+RACERS
+vs
+COPS
+|
+RedlineRP
+Chris
+Stapleton
+-
+“Cold”
+|
+CMA
+Awards
+
+Why
+this
+"falling
+rocks"
+sign
+is
+more
+important
+than
+most
+Scoring
+a
+
+Yard
+Touchdown
+with
+Derrick
+Henry
+on
+Every
+Madden
+Will
+Ferrell
+Shows
+Up
+Instead
+of
+Ryan
+Reynolds
+Boomerang
+Trick
+Shots
+
+|
+Dude
+Perfect
+TODAY
+IS
+MY
+BIRTHDAY
+Mitski
+-
+The
+Only
+Heartbreaker
+(Official
+Video)
+We
+Cleaned
+The
+Beach
+#shorts
+#unfckitchallenge
+#TeamSeas
+It’s
+Always
+Sunny
+in
+Philadelphia
+|
+Season
+:
+Official
+Trailer
+|
+FXX
+Canelo
+Alvarez
+KOs
+Caleb
+Plant
+With
+Crazy
+Uppercut
+|
+SHOWTIME
+PPV
+Pan
+Pizza
+-
+You
+Suck
+at
+Cooking
+(episode-134)
+The
+
+Dollar
+Bento
+Box
+|
+But
+Cheaper
+DIGGING
+A
+SECRET
+TUNNEL
+Part
+
+Hermitcraft
+
+|
+Ep
+:
+Connecting
+to
+the
+Boatem
+GIGA
+Base!
+Moving
+ALL
+My
+Animals
+Into
+The
+Alligator
+House
+🐊🏡
+#shorts
+#animals
+
+Things
+Canelo
+Álvarez
+Can't
+Live
+Without
+|
+GQ
+Sports
+CLIFFORD
+The
+BIG
+RED
+DOG
+In
+GTA
+
+(Movie)
+Rochy
+RD
+-
+APAGANDO
+VELA
+🕯
+x
+Musicologo
+The
+Libro
+|
+Video
+Oficial
+Update
+to
+YouTube's
+Dislike
+Count
+​​​
+🔥
+#shorts
+My
+DIY
+Halloween
+Costumes
+
+|
+Charli
+D’Amelio
+Why
+The
+Golden
+State
+Warriors
+Are
+SO
+HOT
+Right
+Now...
+Lil
+Durk
+-
+Lion
+Eyes
+(Official
+Video)
+Bruno
+Mars,
+Anderson
+Paak,
+Silk
+Sonic
+-
+Smokin
+Out
+The
+Window
+[Official
+Music
+Video]
+Slow
+Mo
+Rainbow
+Fire
+Tornado
+-
+The
+Slow
+Mo
+Guys
+TWICE
+“SCIENTIST”
+M/V
+Teaser
+
+Canelo
+And
+Caleb
+Plant
+Embrace
+And
+Show
+Respect
+After
+Epic
+Battle
+🤝
+I
+Survived
+the
+World's
+Loudest
+Room
+Officially
+Signed
+Michael
+Jordan
+Basketball
+Alex
+Caruso
+On
+What
+Exactly
+Happened
+In
+Free
+Agency
+That
+Led
+Him
+To
+Leave
+The
+Lakers
+For
+The
+Bulls
+I
+WON
+My
+First
+Professional
+Golf
+Tournament
+|
+Micah
+Morris
+Golf
+Richard
+Hammond
+had
+to
+sell
+ALL
+of
+these
+cars
+Ryan
+Reynolds
+Shows
+Up
+Instead
+of
+Will
+Ferrell
+|
+The
+Tonight
+Show
+Starring
+Jimmy
+Fallon
+Polo
+G
+-
+Hall
+of
+Fame
+
+(Album
+Trailer)
+Holiday
+Upgrade
+Can
+Spinlaunch
+Throw
+Rockets
+Into
+Space?
+GTA
+Trilogy
+Definitive
+Edition
+Is
+A
+HUGE
+Disappointment
+BABY'S
+MEETING
+FOR
+THE
+FIRST
+TIME!
+THE
+CUTEST
+REACTION!
+The
+Return
+of
+Jarvis
+When
+the
+game
+gives
+you
+good
+teammates
+Kay’s
+Water
+Broke
+Return
+to
+Island
+of
+the
+Dead!
+-
+DayZ
+Mod
+Jason
+Aldean
+&
+Carrie
+Underwood
+–
+If
+I
+Didn’t
+Love
+You
+(2021
+CMA
+Awards)
+Tate
+McRae
+-
+feel
+like
+shit
+(Official
+Video)
